[flink] PartitionMarkDone enables the use of various partition trigger strategies.
