+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/sink/StoreCommitter.java
@@ -20,7 +20,7 @@ package org.apache.paimon.flink.sink;
 
 import org.apache.paimon.annotation.VisibleForTesting;
 import org.apache.paimon.flink.metrics.FlinkMetricRegistry;
-import org.apache.paimon.flink.sink.partition.PartitionMarkDone;
+import org.apache.paimon.flink.sink.partition.PartitionCollector;
 import org.apache.paimon.io.DataFileMeta;
 import org.apache.paimon.manifest.ManifestCommittable;
 import org.apache.paimon.table.FileStoreTable;
@@ -43,7 +43,7 @@ public class StoreCommitter implements Committer<Committable, ManifestCommittabl
 
     private final TableCommitImpl commit;
     @Nullable private final CommitterMetrics committerMetrics;
-    @Nullable private final PartitionMarkDone partitionMarkDone;
+    private final PartitionCollector partitionCollector;
 
     public StoreCommitter(FileStoreTable table, TableCommit commit, Context context) {
         this.commit = (TableCommitImpl) commit;
@@ -56,8 +56,8 @@ public class StoreCommitter implements Committer<Committable, ManifestCommittabl
         }
 
         try {
-            this.partitionMarkDone =
-                    PartitionMarkDone.create(
+            this.partitionCollector =
+                    PartitionCollector.create(
                             context.streamingCheckpointEnabled(),
                             context.isRestored(),
                             context.stateStore(),
@@ -111,29 +111,23 @@ public class StoreCommitter implements Committer<Committable, ManifestCommittabl
             throws IOException, InterruptedException {
         commit.commitMultiple(committables, false);
         calcNumBytesAndRecordsOut(committables);
-        if (partitionMarkDone != null) {
-            partitionMarkDone.notifyCommittable(committables);
-        }
+        partitionCollector.notifyCommittable(committables);
     }
 
     @Override
     public int filterAndCommit(
             List<ManifestCommittable> globalCommittables, boolean checkAppendFiles) {
         int committed = commit.filterAndCommitMultiple(globalCommittables, checkAppendFiles);
-        if (partitionMarkDone != null) {
-            partitionMarkDone.notifyCommittable(globalCommittables);
-        }
+        partitionCollector.notifyCommittable(globalCommittables);
         return committed;
     }
 
     @Override
     public Map<Long, List<Committable>> groupByCheckpoint(Collection<Committable> committables) {
-        if (partitionMarkDone != null) {
-            try {
-                partitionMarkDone.snapshotState();
-            } catch (Exception e) {
-                throw new RuntimeException(e);
-            }
+        try {
+            partitionCollector.snapshotState();
+        } catch (Exception e) {
+            throw new RuntimeException(e);
         }
 
         Map<Long, List<Committable>> grouped = new HashMap<>();
@@ -146,9 +140,7 @@ public class StoreCommitter implements Committer<Committable, ManifestCommittabl
     @Override
     public void close() throws Exception {
         commit.close();
-        if (partitionMarkDone != null) {
-            partitionMarkDone.close();
-        }
+        partitionCollector.close();
     }
 
     private void calcNumBytesAndRecordsOut(List<ManifestCommittable> committables) {
